Implements handling of meta_* timestamps
Handles timestamp types meta_created and meta_modified, to
search on modification of metadata rather than data itself.

It would be nice if there was a way to search on *any*
update to an object/collection but unfortunately due to
limitations in the iRODS genQuery API, I do not think it
is possible to do that (without an installed SQL query)
in a single iRODS query.

Perhaps baton-metaquery could be made to aggregate the
results of multiple genQuery calls and then uniq them
before populating the AVU/ACL/timestamp fields, but
that has not been included here.