Added operations required by a Go client wrapping baton-do.
Added a remove data object operation.

Added a create and remove collection option (with recursion).

Added the ability to optionally create and return checksums on put.

Ensured that all operations return a JSON document describing the
target, even in cases where the target was deleted. This is for
consistency in the API.