Added a GitHub Action build and test.
This build uses an iRODS 4.1.12 Docker container to supply a test
iRODS instance.

The build takes place in a Docker container set up with Conda, using a
custom local Conda recipe to build from the local source. Access to
the iRODS instance is provided by a Docker network.

GitHub actions appear to override the Docker entrypoint by default, so
a copy of the original entrypoint script is added here to restore the
original default entrypoint of the Conda container.

The build script performs both make check and make distcheck. It would
be nice to separate these, but that would involve having two Conda
recipes.