Explicitly free JSON checksums only on error
This attempts to fix a bug which intermittently affected the client
when performing checksum operations. The symptom would be that no JSON
document would be printed when baton-do dispatched to "checksum".

Investigating this led to finding an error where checksum JSON was
overzealously freed. Fixing that also appeared to fix the bug, but I'm
not 100% sure it is the root cause.

The JSON-ified checksum being added to the object (JSON pointer)
passed in as target to baton_json_checksum_op was being json_free'd
before return. This had the effect of freeing the entire target (?),
which is used subsequently by the caller to print a JSON
document. This meant that the caller would print an empty document.

The JSON checksum should be freed only if it was not possible to
transfer ownership to the target that was passed in.