Add support for iRODS 4.2.9
iRODS 4.2.9 makes some API changes which are accommodated here, while
retaining backward compatibility with previous 4.x versions.

The most significant change is the addition of the ability to select
"good" replicates i.e. at-rest on the server. Previously we were able
only to select "newly created" replicates i.e. those which were not
stale, but which could be in-flight. This change in iRODS was made
possible by adding object locking to the server.

The API features for the old-style "newly created" replicates are
marked as deprecated and redirect to API for "good" replicates on
iRODS >= 4.2.9.

Add default of creating checksums to checksum_data_obj if no operation
is specified.

Add -f option to ichksum for creating test data so that checksums get
created. This is a change to ichksum behaviour in 4.2.9.