Add a regression test for GitHub issue #252
Add a regression test for this segfault, which is known to occur on
iRODS 4.2.7. iRODS 4.2.10 fixes a server bug which prevents this baton
bug being triggered.

Improve the robustness of the iRODS teardown script by adding an
ichmod to ensure all test data are owned (and deletable) by the
current user before attempting to delete.