Fix timeout threading issues
Incorporate suggestions kindly provided by Rob Davies:

Protect the timeout thread flag with the available mutex.

Use a condition variable to handle timeouts and completion better
than a simple sleep.

Clean up with pthread_join on exit, if a thread was created.

Adjust the logging to include NOTICE level for key events. The logging
is a bit excessive, but we can remove it later, when we've confirmed
that all is behaving well.