Handle change to behaviour of iRODS' clientLogin function
Up to (and including) 4.3.0 the second argument to clientLogin could
be an empty string. From 4.3.1, the client library attempts to parse
this argument as JSON, which fails on an empty string.

clientLogin isn't documented, so we don't know what JSON is valid
here. However, passing a null pointer appears to give a usable result,
so doing that for now.

Also fixed a new compile warning in the tests.