Add non-expanded ACL reporting for collections
To date baton reported non-expanded ACLs for data objects only
(expansion meaning an access control for a group reporting all group
members in addition to the group itself). This was because baton
avoided using specific queries in case the required query was not
installed.

This change achieves non-expanded ACLs for collections in the same way
that the native "ils" icommand does; by using a specific query that is
supplied with iRODS and therefore wil be present unless removed by an
administrator.