Fix initialisation of rodsPath_t
As of iRODS 4.3.4 the memory management semantics of getRodsObjType
have changed. It is now required that the rodsObjStat member must be
initialised before calling. Previously the call would initialise this
member. During the call this member is unconditionally freed, so must
be initialised by the caller to avoid a segfault. It's an improvement
to initialise this explicitly, anyway.